move gp_upgrade version functionality to hub architecture
