+++ gpMgmt/go-utils/src/gp_upgrade/cli/commanders/version_test.go
@@ -1,23 +1,24 @@
-package commands
+package commanders_test
 
 import (
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
+	"gp_upgrade/cli/commanders"
 )
 
 var _ = Describe("Version", func() {
 	Describe("VersionString", func() {
 		Context("when global var GpdbVersion is the empty string", func() {
 			It("returns the default version", func() {
-				GpdbVersion = ""
-				Expect(versionString()).To(Equal("gp_upgrade unknown version"))
+				commanders.GpdbVersion = ""
+				Expect(commanders.VersionString()).To(Equal("gp_upgrade unknown version"))
 			})
 		})
 
 		Context("when global var GpdbVersion is set to something", func() {
 			It("returns what it's set to", func() {
-				GpdbVersion = "Something"
-				Expect(versionString()).To(Equal("gp_upgrade version Something"))
+				commanders.GpdbVersion = "Something"
+				Expect(commanders.VersionString()).To(Equal("gp_upgrade version Something"))
 			})
 		})
 	})
